requirements.txt
streamlit>=1.28.0
streamlit-webrtc>=0.47.0
opencv-python>=4.8.0
pyzbar>=0.1.9
pandas>=2.0.0
openpyxl>=3.1.0
av>=10.0.0
numpy>=1.24.0
Pillow>=9.0.0